+++ src/accessory/sensor/contact.ts
@@ -5,6 +5,7 @@ import { SensorAccessory } from './sensor.js';
 import { strings } from '../../i18n/i18n.js';
 
 import { AccessoryType, HKCharacteristicKey } from '../../model/enums.js';
+import { HistoryType } from '../../model/history.js';
 import { ContactSensorConfig, MQTTAccessoryDependency } from '../../model/types.js';
 
 export class ContactSensorAccessory extends SensorAccessory<ContactSensorConfig> {
@@ -24,5 +25,6 @@ export class ContactSensorAccessory extends SensorAccessory<ContactSensorConfig>
   private async onContactStateUpdate(topic: string, value: PrimitiveTypes): Promise<void> {
     const numeric = value === this.getPrimitiveValue('valueContactDetected') ? 0 : 1;
     this.onUpdate(HKCharacteristicKey.ContactSensorState, numeric, numeric ? strings.sensor.contact.inactive : strings.sensor.contact.active);
+    this.recordHistory(HistoryType.DOOR, { status: numeric }, true);
   }
 }